hotfix
Изменения:
- добавлены приписьки* Handler к ручками;
- теперь ошибка правильно обрабатывается ошибка 242;
