--- server/api/application/router/handlers/itemHandlers/getItemsData.js
@@ -1,13 +1,0 @@
-const BaseHandler = require('../../BaseHandler.js');
-
-class GetItemsData extends BaseHandler {
-    constructor(db) {
-        super(db);
-    }
-
-    async execute() {
-        return this.db.getAllItemsData();
-    }
-}
-
-module.exports = GetItemsData;
