+++ server/api/application/router/handlers/itemHandlers/getItemsDataHandler.js
@@ -1,0 +1,13 @@
+const BaseHandler = require('../BaseHandler.js');
+
+class GetItemsDataHandler extends BaseHandler {
+    constructor(db) {
+        super(db);
+    }
+
+    async execute() {
+        return this.db.getAllItemsData();
+    }
+}
+
+module.exports = GetItemsDataHandler;
