--- server/api/application/router/handlers/userHandlers/getRatingTable.js
@@ -1,26 +1,0 @@
-const BaseHandler = require('../../BaseHandler.js');
-
-class GetRatingTable extends BaseHandler {
-    constructor(db) {
-        super(db);
-    }
-    
-    async execute(params) {
-        // Проверка наличия всех необходимых параметров
-        if (!params.token) {
-            return { error: 242 };
-        }
-        
-        // Получение пользака по токенуи
-        const user = await this.db.getUserByToken(params.token);
-        if (!user) {
-            return { error: 705 };
-        }
-        
-        // Получение таблицы рейтинга
-        const ratingTable = await this.db.getRatingTable();
-        return ratingTable;
-    }
-}
-
-module.exports = GetRatingTable;
